# Reduce malloc times to make eb_init_encoder faster