… Source/Lib/Common/Codec/EbPictureBufferDesc.c
… Source/Lib/Common/Codec/EbPictureControlSet.c
… Source/Lib/Common/Codec/EbSystemResourceManager.c